Add advanced config section to the captioner
